unsuccessfully move logging to separate file
